Moved QRClickToolButton class to QRDUtils. Setting an empty name in the rename bookmark dialog resets the name to the EID. (+2 squashed commit)
Squashed commit:

[fef253236] Fixed ambiguous QString conversion on Linux.

[e87494d79] Fixed code formatting.
