+++ qrenderdoc/Code/QRDUtils.h
@@ -35,6 +35,7 @@
 #include <QSharedPointer>
 #include <QSortFilterProxyModel>
 #include <QStyledItemDelegate>
+#include <QToolButton>
 #include "Code/Interface/QRDInterface.h"
 
 #if !defined(RELEASE) && !defined(__FreeBSD__)
@@ -988,3 +989,20 @@ void *AccessWaylandPlatformInterface(const QByteArray &resource, QWindow *window
 
 void UpdateVisibleColumns(rdcstr windowTitle, int columnCount, QHeaderView *header,
                           const QStringList &headers);
+
+// A version of QToolButton that can also handle right clicks
+class QRClickToolButton : public QToolButton
+{
+  Q_OBJECT
+
+public:
+  explicit QRClickToolButton(QWidget *parent = 0);
+
+private slots:
+  void mousePressEvent(QMouseEvent *e);
+
+signals:
+  void rightClicked();
+
+public slots:
+};
